[db] finish the simple models

+++ server/api/driver/driver.model.js
@@ -3,13 +3,13 @@
 export default function(sequelize, DataTypes) {
   return sequelize.define('Driver', {
     _id: {
-      type: DataTypes.INTEGER,
+      type: DataTypes.INTEGER(11),
       allowNull: false,
       primaryKey: true,
       autoIncrement: true
     },
     createdAt: DataTypes.DATE,
     updatedAt: DataTypes.DATE,
-    userId: DataTypes.INTEGER
+    userId: DataTypes.INTEGER(11)
   });
 }
